e2eAIOK/SDA/requirements.txt
pyyaml
sigopt
xgboost
scikit-learn
six